First ReadMe update
This is the update for the plugin's readme file. I had to make this update as practice for the Hello World GitHub Guide to using GitHub. I don't think this will be further updated as tutorials and other articles for use will be available on my website, www.osatoworks.com/projects/valour-wp-gallery